correct linter errors
